refactor: update documentation and API references for TypeScript utilities
- Replaced deprecated functions `genClassMethod` and `genClassProperty` with `genProperty` in documentation and examples.
- Enhanced API documentation for class and object generation, including new examples for `genProperty` and `genMethod`.
- Updated the SQLite database to reflect the latest changes in documentation.
- Improved overall organization and clarity of the documentation for better user experience.
