+++ docs/.docs/config/esm.ts
@@ -10,20 +10,14 @@ genImport('vue', [
   'ref',
   'computed',
   'watch'
-])`,
-    output: () => [
-      knitwork.genImport('vue', ['ref']),
-      knitwork.genImport('vue', ['ref', 'computed', 'watch']),
-    ],
-  },
-  {
-    module: 'esm',
-    label: 'genImport (type)',
-    code: `genImport('@nuxt/utils', ['test'], { type: true })
+])
+genImport('@nuxt/utils', ['test'], { type: true })
 genImport('@nuxt/utils', [
   { name: 'test', as: 'value' }
 ], { type: true })`,
     output: () => [
+      knitwork.genImport('vue', ['ref']),
+      knitwork.genImport('vue', ['ref', 'computed', 'watch']),
       knitwork.genImport('@nuxt/utils', ['test'], { type: true }),
       knitwork.genImport('@nuxt/utils', [{ name: 'test', as: 'value' }], { type: true }),
     ],
@@ -34,18 +28,12 @@ genImport('@nuxt/utils', [
     code: `genDynamicImport('pkg')
 genDynamicImport('pkg', {
   wrapper: true
-})`,
+})
+genDynamicImport('pkg', { type: true })
+genDynamicImport('pkg', { type: true, name: 'foo' })`,
     output: () => [
       knitwork.genDynamicImport('pkg'),
       knitwork.genDynamicImport('pkg', { wrapper: true }),
-    ],
-  },
-  {
-    module: 'esm',
-    label: 'genDynamicImport (type)',
-    code: `genDynamicImport('pkg', { type: true })
-genDynamicImport('pkg', { type: true, name: 'foo' })`,
-    output: () => [
       knitwork.genDynamicImport('pkg', { type: true }),
       knitwork.genDynamicImport('pkg', { type: true, name: 'foo' }),
     ],
@@ -54,34 +42,14 @@ genDynamicImport('pkg', { type: true, name: 'foo' })`,
     module: 'esm',
     label: 'genExport',
     code: `genExport('pkg', ['a', 'b'])
-genExport('utils', ['fn'])`,
+genExport('utils', ['fn'])
+genExport('pkg', '*')
+genExport('pkg', { name: '*', as: 'utils' })`,
     output: () => [
       knitwork.genExport('pkg', ['a', 'b']),
       knitwork.genExport('utils', ['fn']),
-    ],
-  },
-  {
-    module: 'esm',
-    label: 'genExportStar',
-    code: `genExportStar('pkg')
-genExportStar('./utils', {
-  singleQuotes: true
-})`,
-    output: () => [
-      knitwork.genExportStar('pkg'),
-      knitwork.genExportStar('./utils', { singleQuotes: true }),
-    ],
-  },
-  {
-    module: 'esm',
-    label: 'genExportStarAs',
-    code: `genExportStarAs('pkg', 'utils')
-genExportStarAs('./helpers', 'Helpers', {
-  singleQuotes: true
-})`,
-    output: () => [
-      knitwork.genExportStarAs('pkg', 'utils'),
-      knitwork.genExportStarAs('./helpers', 'Helpers', { singleQuotes: true }),
+      knitwork.genExport('pkg', '*'),
+      knitwork.genExport('pkg', { name: '*', as: 'utils' }),
     ],
   },
   {
