+++ docs/2.apis/18.object.md
@@ -1,5 +1,5 @@
 ---
-description: Object and collection literal generation (genLiteral, genObject, genArray, genMap, genSet)
+description: Object and collection literal generation (genObject, genArray, genMap, genSet)
 icon: i-lucide-square-function
 ---
 
@@ -22,17 +22,6 @@ Values are not escaped or quoted.
 ```js
 genArray([1, 2, 3])
 // ~> `[1, 2, 3]`
-```
-
-### `genLiteral(fields, indent, _options)`
-
-Create object literal from field descriptors.
-
-**Example:**
-
-```js
-genLiteral(['type', ['type', 'A'], ['...', 'b']])
-// ~> `{ type, type: A, ...b }`
 ```
 
 ### `genMap(entries, indent, options)`
@@ -72,6 +61,56 @@ Values are escaped and quoted if necessary (strings, etc.).
 ```js
 genSet(["foo", "bar", 1]);
 // ~> `new Set(["foo", "bar", 1])`
+```
+
+## Typescript
+
+### `genGetter(name, body, options, indent)`
+
+Generate getter: `get name() { ... }` (for class or object literal).
+
+**Example:**
+
+```js
+genGetter("value", ["return this._v;"]);
+// ~> `get value() { return this._v; }`
+
+genGetter("id", ["return this._id;"], { returnType: "string" });
+// ~> `get id(): string { return this._id; }`
+```
+
+### `genMethod(options, indent)`
+
+Generate method (including get/set) with optional async/generator/static. For class or object literal: `name(params) { body }`, `get name() { }`, `set name(v) { }`.
+
+**Example:**
+
+```js
+genMethod({ name: "foo" });
+// ~> `foo() {}`
+
+genMethod({ name: "bar", parameters: [{ name: "x", type: "string" }], body: ["return x;"], returnType: "string" });
+// ~> `bar(x: string): string { return x; }`
+
+genMethod({ name: "value", kind: "get", body: ["return this._v;"], returnType: "number" });
+// ~> `get value(): number { return this._v; }`
+
+genMethod({ name: "value", kind: "set", parameters: [{ name: "v", type: "number" }], body: ["this._v = v;"] });
+// ~> `set value(v: number) { this._v = v; }`
+```
+
+### `genSetter(name, paramName, body, options, indent)`
+
+Generate setter: `set name(param) { ... }` (for class or object literal).
+
+**Example:**
+
+```js
+genSetter("value", "v", ["this._v = v;"]);
+// ~> `set value(v) { this._v = v; }`
+
+genSetter("id", "x", ["this._id = x;"], { paramType: "string" });
+// ~> `set id(x: string) { this._id = x; }`
 ```
 
 <!-- /automd -->
