+++ docs/2.apis/19.utils.md
@@ -1,5 +1,5 @@
 ---
-description: Utility functions (genRegExp, genJSDocComment, wrapInDelimiters, genKey, genComment)
+description: Utility functions (genLiteral, genRegExp, genJSDocComment, wrapInDelimiters, genKey, genComment)
 icon: i-lucide-square-function
 ---
 
@@ -66,6 +66,17 @@ genKey("foo-bar");
 
 genKey("with space");
 // ~> `"with space"`
+```
+
+### `genLiteral(fields, indent, _options)`
+
+Create object literal from field descriptors.
+
+**Example:**
+
+```js
+genLiteral(['type', ['type', 'A'], ['...', 'b']])
+// ~> `{ type, type: A, ...b }`
 ```
 
 ### `genRegExp(pattern, flags?)`
