+++ docs/2.apis/3.esm.md
@@ -62,42 +62,11 @@ genExport("pkg", "foo");
 genExport("pkg", ["a", "b"]);
 // ~> `export { a, b } from "pkg";`
 
-genExport("pkg", { name: "*", as: "bar" });
-// ~> `export * as bar from "pkg";`
-```
-
-### `genExportStar(specifier, options)`
-
-Generate an ESM `export *` statement (re-export all).
-
-**Example:**
-
-```js
-genExportStar("pkg");
+genExport("pkg", "*");
 // ~> `export * from "pkg";`
 
-genExportStar("./utils", { singleQuotes: true });
-// ~> `export * from './utils';`
-
-genExportStar("pkg", { attributes: { type: "json" } });
-// ~> `export * from "pkg" with { type: "json" };`
-```
-
-### `genExportStarAs(specifier, namespace, options)`
-
-Generate an ESM `export * as` statement (re-export all as namespace).
-
-**Example:**
-
-```js
-genExportStarAs("pkg", "utils");
-// ~> `export * as utils from "pkg";`
-
-genExportStarAs("./helpers", "Helpers", { singleQuotes: true });
-// ~> `export * as Helpers from './helpers';`
-
-genExportStarAs("pkg", "ns", { attributes: { type: "json" } });
-// ~> `export * as ns from "pkg" with { type: "json" };`
+genExport("pkg", { name: "*", as: "bar" });
+// ~> `export * as bar from "pkg";`
 ```
 
 ### `genImport(specifier, imports?, options)`
