+++ docs/2.apis/4.class.md
@@ -1,5 +1,5 @@
 ---
-description: Class and class member generation (genClass, genConstructor, genClassProperty, genClassMethod, get/set)
+description: Class and class member generation (genClass, genConstructor, genProperty, genMethod, get/set)
 icon: i-lucide-square-function
 ---
 
@@ -31,46 +31,6 @@ genClass("Exported", [], { export: true });
 // ~> `export class Exported {}`
 ```
 
-### `genClassMethod(options, indent)`
-
-Generate class method (including get/set) with optional async/generator.
-
-**Example:**
-
-```js
-genClassMethod({ name: "foo" });
-// ~> `foo() {}`
-
-genClassMethod({ name: "bar", parameters: [{ name: "x", type: "string" }], body: ["return x;"], returnType: "string" });
-// ~> `bar(x: string): string { return x; }`
-
-genClassMethod({ name: "value", kind: "get", body: ["return this._v;"], returnType: "number" });
-// ~> `get value(): number { return this._v; }`
-
-genClassMethod({ name: "value", kind: "set", parameters: [{ name: "v", type: "number" }], body: ["this._v = v;"] });
-// ~> `set value(v: number) { this._v = v; }`
-```
-
-### `genClassProperty(name, options, indent)`
-
-Generate class property: `name: Type` or `name = value` (with optional modifiers).
-
-**Example:**
-
-```js
-genClassProperty("x", { type: "number" });
-// ~> `x: number`
-
-genClassProperty("y", { value: "0" });
-// ~> `y = 0`
-
-genClassProperty("z", { type: "string", value: "'z'" });
-// ~> `z: string = 'z'`
-
-genClassProperty("id", { type: "string", readonly: true, static: true });
-// ~> `static readonly id: string`
-```
-
 ### `genConstructor(parameters, body, options, indent)`
 
 Generate constructor(params) { [super(...);] ... }.
@@ -88,32 +48,27 @@ genConstructor([{ name: "a", type: "number" }, { name: "b", type: "number" }], [
 // ~> `constructor(a: number, b: number) { super(a, b); }`
 ```
 
-### `genGetter(name, body, options, indent)`
+### `genProperty(field, indent)`
 
-Generate getter: `get name() { ... }` (for class or object literal).
-
-**Example:**
-
-```js
-genGetter("value", ["return this._v;"]);
-// ~> `get value() { return this._v; }`
-
-genGetter("id", ["return this._id;"], { returnType: "string" });
-// ~> `get id(): string { return this._id; }`
-```
-
-### `genSetter(name, paramName, body, options, indent)`
-
-Generate setter: `set name(param) { ... }` (for class or object literal).
+Generate a single property signature from a TypeField. Returns `[modifiers?][name][optional?]: [type][ = value]?`. When `field.jsdoc` is set, prepends JSDoc comment. Supports static, readonly, public, private, protected (class property) and value (initializer).
 
 **Example:**
 
 ```js
-genSetter("value", "v", ["this._v = v;"]);
-// ~> `set value(v) { this._v = v; }`
+genProperty({ name: "foo", type: "string" });
+// ~> `foo: string`
 
-genSetter("id", "x", ["this._id = x;"], { paramType: "string" });
-// ~> `set id(x: string) { this._id = x; }`
+genProperty({ name: "bar", type: "number", optional: true });
+// ~> `bar?: number`
+
+genProperty({ name: "id", type: "string", jsdoc: "Unique id" }, "  ");
+// ~> `/** Unique id *\/\n  id: string`
+
+genProperty({ name: "x", value: "0" });
+// ~> `x = 0`
+
+genProperty({ name: "id", type: "string", readonly: true, static: true });
+// ~> `static readonly id: string`
 ```
 
 <!-- /automd -->
