+++ docs/2.apis/7.function.md
@@ -1,9 +1,9 @@
 ---
-description: Function and block generation (genFunction, genArrowFunction, genMethod, genBlock, genParam)
+description: Function and block generation (genFunction, genArrowFunction, genBlock, genParam)
 icon: i-lucide-square-function
 ---
 
-This module provides APIs for generating function declarations, arrow functions, object methods, statement blocks, and parameter lists, with support for generics, async, generator, and JSDoc.
+This module provides APIs for generating function declarations, arrow functions, statement blocks, and parameter lists, with support for generics, async, generator, and JSDoc. For object/class methods (genMethod, genGetter, genSetter), see the object module.
 
 <example module="function"></example>
 
@@ -78,26 +78,6 @@ genFunction({ name: "id", generics: [{ name: "T" }], parameters: [{ name: "x", t
 
 genFunction({ name: "foo", export: true });
 // ~> `export function foo() {}`
-```
-
-### `genMethod(options, indent)`
-
-Generate shorthand object method: `name(params) { body }` (no `function` keyword).
-
-**Example:**
-
-```js
-genMethod({ name: "foo" });
-// ~> `foo() {}`
-
-genMethod({ name: "bar", parameters: [{ name: "x", type: "string" }], body: ["return x;"] });
-// ~> `bar(x: string) {\n  return x;\n}`
-
-genMethod({ name: "add", parameters: [{ name: "a", type: "number" }, { name: "b", type: "number" }], returnType: "number", body: ["return a + b;"] });
-// ~> `add(a: number, b: number): number {\n  return a + b;\n}`
-
-genMethod({ name: "fetch", async: true, parameters: [{ name: "url", type: "string" }], body: ["return await fetch(url);"] });
-// ~> `async fetch(url: string) {\n  return await fetch(url);\n}`
 ```
 
 ### `genParam(p)`
