+++ docs/2.apis/8.type.md
@@ -5,9 +5,9 @@ icon: i-lucide-square-function
 
 This module provides APIs for generating type aliases, object types, union/intersection types, mapped types, template literal types, and keyof/typeof/assertion/satisfies.
 
-<example module="type-alias"></example>
+<example module="type"></example>
 
-<!-- automd:jsdocs src="../../src/type-alias.ts" -->
+<!-- automd:jsdocs src="../../src/type.ts" -->
 
 ## Typescript
 
@@ -68,23 +68,6 @@ genMappedType("K", "keyof T", "U");
 
 genMappedType("P", "keyof T", "T[P]");
 // ~> `{ [P in keyof T]: T[P] }`
-```
-
-### `genProperty(field, indent)`
-
-Generate a single property signature from a TypeField. Returns `[name][optional?]: [type]`. When `field.jsdoc` is set, prepends JSDoc comment.
-
-**Example:**
-
-```js
-genProperty({ name: "foo", type: "string" });
-// ~> `foo: string`
-
-genProperty({ name: "bar", type: "number", optional: true });
-// ~> `bar?: number`
-
-genProperty({ name: "id", type: "string", jsdoc: "Unique id" }, "  ");
-// ~> `/** Unique id *\/\n  id: string`
 ```
 
 ### `genSatisfies(expr, type)`
