+++ src/class.ts
@@ -1,14 +1,9 @@
-import { genJSDocComment, wrapInDelimiters } from "./utils";
+import { genJSDocComment, genKey, wrapInDelimiters } from "./utils";
 import { genBlock, genParam } from "./function";
 import type {
   TypeField,
-  TypeGeneric,
   GenClassOptions,
   GenConstructorOptions,
-  GenClassPropertyOptions,
-  GenClassMethodOptions,
-  GenGetterOptions,
-  GenSetterOptions,
 } from "./types";
 
 /**
@@ -115,210 +110,58 @@ export function genConstructor(
 }
 
 /**
- * Generate class property: `name: Type` or `name = value` (with optional modifiers).
+ * Generate a single property signature from a TypeField.
+ * Returns `[modifiers?][name][optional?]: [type][ = value]?`. When `field.jsdoc` is set, prepends JSDoc comment.
+ * Supports static, readonly, public, private, protected (class property) and value (initializer).
  *
  * @example
  *
  * ```js
- * genClassProperty("x", { type: "number" });
- * // ~> `x: number`
+ * genProperty({ name: "foo", type: "string" });
+ * // ~> `foo: string`
  *
- * genClassProperty("y", { value: "0" });
- * // ~> `y = 0`
+ * genProperty({ name: "bar", type: "number", optional: true });
+ * // ~> `bar?: number`
  *
- * genClassProperty("z", { type: "string", value: "'z'" });
- * // ~> `z: string = 'z'`
+ * genProperty({ name: "id", type: "string", jsdoc: "Unique id" }, "  ");
+ * // ~> `/** Unique id *\/\n  id: string`
  *
- * genClassProperty("id", { type: "string", readonly: true, static: true });
+ * genProperty({ name: "x", value: "0" });
+ * // ~> `x = 0`
+ *
+ * genProperty({ name: "id", type: "string", readonly: true, static: true });
  * // ~> `static readonly id: string`
  * ```
  *
- * @param name - property name
- * @param options - type, value, static, readonly, public, private, protected, optional
+ * @param field - property field (name, type, optional, value, modifiers, jsdoc)
  * @param indent - base indent
  * @group Typescript
  */
-export function genClassProperty(
-  name: string,
-  options: GenClassPropertyOptions = {},
-  indent = "",
-): string {
-  const {
-    type,
-    value,
-    static: isStatic,
-    readonly,
-    public: isPublic,
-    private: isPrivate,
-    protected: isProtected,
-    optional,
-  } = options;
-
+export function genProperty(field: TypeField, indent = ""): string {
   const mods = [
-    isStatic && "static",
-    readonly && "readonly",
-    isPublic && "public",
-    isPrivate && "private",
-    isProtected && "protected",
+    field.static && "static",
+    field.readonly && "readonly",
+    field.public && "public",
+    field.private && "private",
+    field.protected && "protected",
   ].filter(Boolean);
-
   const modPart = mods.length > 0 ? mods.join(" ") + " " : "";
-  const opt = optional ? "?" : "";
+  const key = genKey(field.name);
+  const opt = field.optional ? "?" : "";
   let decl: string;
-  if (value === undefined) {
-    decl = type ? `${name}${opt}: ${type}` : name + opt;
+  if (field.value === undefined) {
+    decl = field.type ? `${key}${opt}: ${field.type}` : key + opt;
   } else {
-    decl = type
-      ? `${name}${opt}: ${type} = ${value}`
-      : `${name}${opt} = ${value}`;
+    decl = field.type
+      ? `${key}${opt}: ${field.type} = ${field.value}`
+      : `${key}${opt} = ${field.value}`;
   }
-  return `${indent}${modPart}${decl}`;
-}
-
-/**
- * Generate class method (including get/set) with optional async/generator.
- *
- * @example
- *
- * ```js
- * genClassMethod({ name: "foo" });
- * // ~> `foo() {}`
- *
- * genClassMethod({ name: "bar", parameters: [{ name: "x", type: "string" }], body: ["return x;"], returnType: "string" });
- * // ~> `bar(x: string): string { return x; }`
- *
- * genClassMethod({ name: "value", kind: "get", body: ["return this._v;"], returnType: "number" });
- * // ~> `get value(): number { return this._v; }`
- *
- * genClassMethod({ name: "value", kind: "set", parameters: [{ name: "v", type: "number" }], body: ["this._v = v;"] });
- * // ~> `set value(v: number) { this._v = v; }`
- * ```
- *
- * @param options - name, parameters, body, kind (method/get/set), static, async, generator, returnType, generics, jsdoc
- * @param indent - base indent
- * @group Typescript
- */
-export function genClassMethod(
-  options: GenClassMethodOptions & { name: string },
-  indent = "",
-): string {
-  const {
-    name,
-    parameters = [],
-    body = [],
-    kind = "method",
-    static: isStatic,
-    async: isAsync,
-    generator: isGenerator,
-    returnType,
-    generics = [],
-    jsdoc,
-  } = options;
-
-  const jsdocComment = jsdoc === undefined ? "" : genJSDocComment(jsdoc);
-
-  const genericPart =
-    generics.length > 0
-      ? "<" +
-        generics
-          .map((g: TypeGeneric) => {
-            let s = g.name;
-            if (g.extends) s += ` extends ${g.extends}`;
-            if (g.default) s += ` = ${g.default}`;
-            return s;
-          })
-          .join(", ") +
-        ">"
-      : "";
-
-  const paramsPart = "(" + parameters.map((p) => genParam(p)).join(", ") + ")";
-  const returnPart = returnType ? `: ${returnType}` : "";
-  const bodyContent = genBlock(body.length > 0 ? body : undefined, indent);
-
-  let prefix: string;
-  if (kind === "get") {
-    prefix = [isStatic && "static", "get", name + "()" + returnPart]
-      .filter(Boolean)
-      .join(" ");
-  } else if (kind === "set") {
-    prefix = [isStatic && "static", "set", name + paramsPart]
-      .filter(Boolean)
-      .join(" ");
-  } else {
-    const namePart =
-      name + genericPart + (isGenerator ? "*" : "") + paramsPart + returnPart;
-    prefix = [isStatic && "static", isAsync && "async", namePart]
-      .filter(Boolean)
-      .join(" ");
+  const propLine = modPart + decl;
+  if (field.jsdoc === undefined) {
+    return indent ? `${indent}${propLine}` : propLine;
   }
-
-  return `${jsdocComment}${indent}${prefix} ${bodyContent}`;
-}
-
-/**
- * Generate getter: `get name() { ... }` (for class or object literal).
- *
- * @example
- *
- * ```js
- * genGetter("value", ["return this._v;"]);
- * // ~> `get value() { return this._v; }`
- *
- * genGetter("id", ["return this._id;"], { returnType: "string" });
- * // ~> `get id(): string { return this._id; }`
- * ```
- *
- * @param name - getter name
- * @param body - getter body statements
- * @param options - returnType, jsdoc
- * @param indent - base indent
- * @group Typescript
- */
-export function genGetter(
-  name: string,
-  body: string[] = [],
-  options: GenGetterOptions = {},
-  indent = "",
-): string {
-  const returnPart = options.returnType ? `: ${options.returnType}` : "";
-  const jsdocComment =
-    options.jsdoc === undefined ? "" : genJSDocComment(options.jsdoc);
-  const block = genBlock(body.length > 0 ? body : undefined, indent);
-  return `${jsdocComment}${indent}get ${name}()${returnPart} ${block}`;
-}
-
-/**
- * Generate setter: `set name(param) { ... }` (for class or object literal).
- *
- * @example
- *
- * ```js
- * genSetter("value", "v", ["this._v = v;"]);
- * // ~> `set value(v) { this._v = v; }`
- *
- * genSetter("id", "x", ["this._id = x;"], { paramType: "string" });
- * // ~> `set id(x: string) { this._id = x; }`
- * ```
- *
- * @param name - setter name
- * @param paramName - parameter name for the setter
- * @param body - setter body statements
- * @param options - paramType, jsdoc
- * @param indent - base indent
- * @group Typescript
- */
-export function genSetter(
-  name: string,
-  paramName: string,
-  body: string[] = [],
-  options: GenSetterOptions = {},
-  indent = "",
-): string {
-  const paramPart = options.paramType
-    ? `${paramName}: ${options.paramType}`
-    : paramName;
-  const jsdocComment =
-    options.jsdoc === undefined ? "" : genJSDocComment(options.jsdoc);
-  const block = genBlock(body.length > 0 ? body : undefined, indent);
-  return `${jsdocComment}${indent}set ${name}(${paramPart}) ${block}`;
+  const jsdocComment = genJSDocComment(field.jsdoc, indent);
+  return indent
+    ? `${jsdocComment}${indent}${propLine}`
+    : `${jsdocComment}${propLine}`;
 }
